Add a next command
While a task's name can be anything you can use for a file name,
I like to use a number, like 0027, for example.  It's just as
meaningful as anything else.  However, I don't always remember
what number I used last, so add the 'next' command to go figure
it out for me used a "lastnum" file in .dbs and looking for files/
tasks already using the lastnum to figure out what the next number
should be.

Signed-off-by: Al Stone <ahs3@redhat.com>